+++ src/remarkable-pro/components/charts/tables/HeatMapPro/HeatMapPro.emb.ts
@@ -1,44 +1,26 @@
 import { loadData } from '@embeddable.com/core';
 import { defineComponent, EmbeddedComponentMeta, Inputs } from '@embeddable.com/react';
-import PivotTablePro from './index';
+import HeatMapPro from './index';
 import {
   dataset,
   description,
   title,
   maxResults,
-  measures,
-  genericBoolean,
-  subInputGenericBoolean,
   genericString,
   dimensionWithDateBounds,
   genericNumber,
+  measure,
+  inputColor,
+  genericBoolean,
 } from '../../../component.constants';
 
 export const meta = {
-  name: 'PivotTablePro',
-  label: 'Pivot Table',
+  name: 'HeatMapPro',
+  label: 'Heat Map',
   category: 'Table Charts',
   inputs: [
     dataset,
-    {
-      ...measures,
-      label: 'Measures To Display',
-      inputs: [
-        ...measures.inputs,
-        {
-          ...subInputGenericBoolean,
-          name: 'showColumnTotal',
-          label: 'Show Column Total',
-        },
-        {
-          ...subInputGenericBoolean,
-          name: 'showAsPercentage',
-          label: 'Show As Percentage',
-          description: 'If turned on, other measures may be ignored',
-          defaultValue: false,
-        },
-      ],
-    },
+    measure,
     {
       ...dimensionWithDateBounds,
       label: 'Row Dimension',
@@ -51,8 +33,23 @@ export const meta = {
     },
     title,
     description,
-    { ...genericBoolean, name: 'showRowTotals', label: 'Show Row Totals' },
     { ...genericString, name: 'displayNullAs', label: 'Display Null As' },
+    { ...inputColor, name: 'maxColor', label: 'Max Color', defaultValue: 'green', required: true },
+    { ...inputColor, name: 'midColor', label: 'Mid Color', defaultValue: 'yellow', required: true },
+    { ...inputColor, name: 'minColor', label: 'Min Color', defaultValue: 'red', required: true },
+    {
+      ...genericString,
+      name: 'minThreshold',
+      label: 'Max range lower limit',
+      description: 'Enter a value as either a number (e.g. 20) or a percentage (e.g. 20%)',
+    },
+    {
+      ...genericString,
+      name: 'maxThreshold',
+      label: 'Min range upper limit',
+      description: 'Enter a value as either a number (e.g. 20) or a percentage (e.g. 20%)',
+    },
+    { ...genericBoolean, name: 'showValues', label: 'Show Values', defaultValue: true },
     {
       ...genericNumber,
       name: 'firstColumnWidth',
@@ -65,19 +62,17 @@ export const meta = {
       label: 'Column Width',
       description: 'Set the width in px (e.g. 200)',
     },
-
     maxResults,
   ],
 } as const satisfies EmbeddedComponentMeta;
 
-export default defineComponent(PivotTablePro, meta, {
+export default defineComponent(HeatMapPro, meta, {
   props: (inputs: Inputs<typeof meta>) => {
     return {
       ...inputs,
-
       results: loadData({
         from: inputs.dataset,
-        select: [inputs.rowDimension, inputs.columnDimension, ...inputs.measures],
+        select: [inputs.rowDimension, inputs.columnDimension, inputs.measure],
         limit: inputs.maxResults,
         countRows: true,
       }),