+++ src/remarkable-ui/charts/tables/PivotTable/PivotTable.types.tsx
@@ -38,4 +38,5 @@ export type PivotTableProps<T> = {
   percentageDecimalPlaces?: number;
   columnWidth?: number;
   firstColumnWidth?: number;
+  className?: string;
 };