+++ src/remarkable-ui/index.ts
@@ -30,15 +30,24 @@ export { TextField } from './editors/TextField/TextField';
 export { NumberField } from './editors/NumberField/NumberField';
 
 // Charts
+export { BarChart } from './charts/bars/BarChart';
+export { getBarChartData, getBarChartOptions } from './charts/bars/bars.utils';
 export { LineChart } from './charts/lines/LineChart';
+export { getLineChartData, getLineChartOptions } from './charts/lines/lines.utils';
 export { KpiChart } from './charts/kpis/KpiChart';
 export { DonutChart } from './charts/pies/DonutChart';
 export { PieChart } from './charts/pies/PieChart';
-export { defaultPieChartOptions } from './charts/pies/pies.constants';
+export {
+  defaultPieChartOptions,
+  defaultDonutChartOptions,
+  defaultDonutLabelChartOptions,
+} from './charts/pies/pies.constants';
 export { TablePaginated } from './charts/tables/Table/TablePaginated';
 export * from './charts/tables/Table/table.types';
 export { PivotTable } from './charts/tables/PivotTable/PivotTable';
 export * from './charts/tables/PivotTable/PivotTable.types';
+export { HeatMap } from './charts/tables/HeatMap/HeatMap';
+export * from './charts/tables/HeatMap/HeatMap.types';
 
 // Chart Utils
 export { getTableTotalPages } from './charts/tables/Table/components/TablePagination/TablePagination';